Removes the copy of /neille from /docs/arch/fig and restores the symbolic link
As a result of a previous arbitrary copying of / prior to the initial import, a hard copy of /neille was noticed to reside under /docs/arch/fig.

Originally, it was intended for a symbolic link under /docs/arch/fig to point to the source code of Neille.

The extra copy of /neille is now removed and the original symbolic link is restored.