Changes the way card graphics are communicated to the program in Racket
Instead of generating thunks that would directly construct instances of the bitmap% object after application, card images are now encoded as an instance of the meta-image structure.

The program in Racket reads the image metadata from instances of the meta-image structure in order to construct bitmap% instances for the required geometries of a given card.

Note that, bitmap% instances are not directly loaded into memory. A promise is however generated to do so when required at runtime.

The changes in this commit are intended to make instances of the ws-card structure serializable.

Thanks to Eli Barzilay
http://racket-lang.org/irc-logs/20110830.txt