Implements persistent storage of players' data
Most changes in this commit are necessary to ensure the serializability
of objects instances that are wished to become stored on hard disk and
resurrected during the next run of the program.  The implementation of
cardlist based on a two-way linked list is also removed in this commit.
Our goal is to continue with the implementation based on plain Racket
lists.